ai-evals/pyproject.toml
[project]
name = "fit-ai-evals"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "pandas>=2.2.0",
  "python-dotenv>=1.0.1",
  "ragas>=0.4.3",
  "langchain-openai>=0.3.35",
  "supabase>=2.28.0",
  "voyageai>=0.3.7",
]

[tool.uv]
package = false
